backend/requirements.txt
rdflib>=7.0,<8.0
numpy>=2.0
opencv-python>=4.10
Pillow>=10.0
networkx>=3.3
requests>=2.32
tqdm>=4.66
pytesseract>=0.3.10
python-louvain>=0.16
fastapi>=0.115
uvicorn[standard]>=0.32
python-dotenv>=1.0
